+++ core.impl/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/impl/repair/state/ReplicationStateImpl.java
@@ -250,6 +250,6 @@ public class ReplicationStateImpl implements ReplicationState
         // Assuming murmur3 partitioner
         long start = ((Murmur3Token) range.getStart()).getValue();
         long end = ((Murmur3Token) range.getEnd()).getValue();
-        return new LongTokenRange(start, end);
+        return LongTokenRange.of(start, end);
     }
 }